requirements.txt
tensorflow
notebook
matplotlib
numpy
pillow
tensorflow_hub